Use MUnit in ResponseGeneratorSuite
